+++ Exercicios/ex11.py
@@ -1,8 +1,8 @@
 #Faça um programa que leia o preço de um produto e retorne
 #o seu preço com 5%de desconto 
 
-produto = int(input("Quanto custa o produto: "))
-q_desconto = int(input("Desconto de:  "))
+produto = float(input("Quanto custa o produto: "))
+q_desconto = float(input("Desconto de:  "))
 valor_produto = (produto*q_desconto/100)
 
 print(
