+++ Exercicios/ex12.py
@@ -1,5 +1,5 @@
 #calcule um almento de 15% no salario de um funcionario
-salario = int(input("Qual eo seu salario: "))
+salario = float(input("Qual eo seu salario: "))
 salario_almento= (salario*1.15)
 almento = (salario_almento-salario)
